Cleaned up Agenda custom commands.
Cleaned things up and made a stab at getting it to work the way I
want. In particular, I wanted my primary agenda view (C-c a j) to have a
list of PROJECT's w/out the corresponding TODO's interleaved. This work
seemed to hint that my whole custom cmd setup was not formatted
correctly. Amazing that it worked at all.